format script: add MultiROM compatibility
When installing a ROM with this script inside, /system is checked and formatted depending on which filesystem is found. If we install as secondary through MultiROM system partition is still formatted because this script is executed, breaking both primary ROM and MultiROM because this last can't be used if there isn't a primary ROM inside the device. We don't want this behavior so with this update if the primary ROM is being installed/upgraded run filesystem check and format, else don't run it.