Because fuck Google
- Disable package verifier by default

With March's security updates, Google updated their Google services and now
this is getting in the way of Titanium backup and Substratum. Two things that
our users use A LOT!

If anyone is scared that the Boogeyman is going to get them, don't download
pirated apps and you'll be good! :-)

Change-Id: I025e4f599c1fafea12b3471c85b476e7b0a2afa7
Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>